# frontend/src/app/openapi/openapi.json
{"openapi":"3.0.1","info":{"title":"OpenApi specification - Amine","description":"OpenApi documentation for Spring Security","termsOfService":"Terms of service","license":{"name":"Licence name","url":"https://some-url.com"},"version":"1.0"},"servers":[{"url":"http://localhost:8081/api/v1","description":"Local ENV"}],"security":[{"bearerAuth":[]}],"paths":{"/users/{id}":{"get":{"tags":["user-controller"],"operationId":"getUserById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}}},"put":{"tags":["user-controller"],"operationId":"updateUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}}}},"/admin/unban/{id}":{"put":{"tags":["admin-controller"],"operationId":"unbanUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"string"}}}}}}},"/admin/ban/{id}":{"put":{"tags":["admin-controller"],"operationId":"banUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"string"}}}}}}},"/users/{id}/upload":{"post":{"tags":["user-controller"],"operationId":"uploadProfilePicture","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"required":["file"],"type":"object","properties":{"file":{"type":"string","format":"binary"}}}}}},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"string"}}}}}}},"/auth/reset-password":{"post":{"tags":["Authentication"],"operationId":"resetPassword","requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}}}}},"/auth/request-password-reset":{"post":{"tags":["Authentication"],"operationId":"requestPasswordReset","parameters":[{"name":"email","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"string"}}}}}}},"/auth/register":{"post":{"tags":["Authentication"],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegistrationRequest"}}},"required":true},"responses":{"202":{"description":"Accepted","content":{"application/json":{"schema":{"type":"string"}}}}}}},"/auth/refresh-token":{"post":{"tags":["Authentication"],"operationId":"refreshToken","requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthenticationResponse"}}}}}}},"/auth/authenticate":{"post":{"tags":["Authentication"],"operationId":"authenticate","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthenticationRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthenticationResponse"}}}}}}},"/admin/request-unban":{"post":{"tags":["admin-controller"],"operationId":"requestUnban","parameters":[{"name":"email","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"string"}}}}}}},"/auth/activate-account":{"get":{"tags":["Authentication"],"operationId":"confirm","parameters":[{"name":"token","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"string"}}}}}}},"/admin/list_user":{"get":{"tags":["admin-controller"],"operationId":"getAllUsers","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/User"}}}}}}}},"/admin/delete/{id}":{"delete":{"tags":["admin-controller"],"operationId":"deleteUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"string"}}}}}}}},"components":{"schemas":{"GrantedAuthority":{"type":"object","properties":{"authority":{"type":"string"}}},"Role":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"createdDate":{"type":"string","format":"date-time"},"lastModifiedDate":{"type":"string","format":"date-time"}}},"User":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"firstname":{"type":"string"},"lastname":{"type":"string"},"dateOfBirth":{"type":"string","format":"date"},"phoneNumber":{"type":"string"},"gender":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"accountLocked":{"type":"boolean"},"enabled":{"type":"boolean"},"roles":{"type":"array","items":{"$ref":"#/components/schemas/Role"}},"profilePicUrl":{"type":"string"},"createdDate":{"type":"string","format":"date-time"},"lastModifiedDate":{"type":"string","format":"date-time"},"name":{"type":"string"},"username":{"type":"string"},"authorities":{"type":"array","items":{"$ref":"#/components/schemas/GrantedAuthority"}},"accountNonExpired":{"type":"boolean"},"credentialsNonExpired":{"type":"boolean"},"accountNonLocked":{"type":"boolean"}}},"RegistrationRequest":{"required":["email","firstname","lastname","password"],"type":"object","properties":{"firstname":{"type":"string"},"lastname":{"type":"string"},"email":{"type":"string"},"password":{"maxLength":2147483647,"minLength":8,"type":"string"}}},"AuthenticationResponse":{"type":"object","properties":{"token":{"type":"string"},"refreshToken":{"type":"string"}}},"AuthenticationRequest":{"required":["email","password"],"type":"object","properties":{"email":{"type":"string"},"password":{"maxLength":2147483647,"minLength":8,"type":"string"}}}},"securitySchemes":{"bearerAuth":{"type":"oauth2","description":"JWT auth description","in":"header","scheme":"bearer","bearerFormat":"JWT"}}}}